requirements.txt
django_mysql
xhtml2pdf
xlrd
pycurl
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
idna==2.8
image==1.5.27
numpy==1.16.2
pandas==0.24.2
pdfkit==0.6.1
Pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
